Moved all source files to icu folder
