--- can.cpp
@@ -1,142 +1,0 @@
-#include "can.h"
-#include "config.h"
-
-//——————————————————————————————————————————————————————————————————————————————
-
-#include <ACAN2515.h>
-
-//——————————————————————————————————————————————————————————————————————————————
-// The Pico has two SPI peripherals, SPI and SPI1. Either (or both) can be used.
-// The are no default pin assignments so they must be set explicitly.
-// Testing was done with Earle Philhower's arduino-pico core:
-// https://github.com/earlephilhower/arduino-pico
-//——————————————————————————————————————————————————————————————————————————————
-
-// This code is just the ACAN2515 LoopBackDemoRaspberryPiPico example code
-
-// Using SPI
-static const byte PICO_SCK_MCP2515  = 18; // SCK input of MCP2515
-static const byte PICO_MOSI_MCP2515 = 19; // SDI input of MCP2515
-static const byte PICO_MISO_MCP2515 = 16; // SDO output of MCP2517
-
-static const byte PICO_CS_MCP2515  = 17;  // CS input of MCP2515 (adapt to your design)
-static const byte PICO_INT_MCP2515 = 15;  // INT output of MCP2515 (adapt to your design)
-
-//——————————————————————————————————————————————————————————————————————————————
-//  MCP2515 Driver object
-//——————————————————————————————————————————————————————————————————————————————
-
-ACAN2515 can (PICO_CS_MCP2515, SPI, PICO_INT_MCP2515) ;
-
-//——————————————————————————————————————————————————————————————————————————————
-//  MCP2515 Quartz: adapt to your design
-//——————————————————————————————————————————————————————————————————————————————
-
-static const uint32_t QUARTZ_FREQUENCY = 16UL * 1000UL * 1000UL; // 16 MHz
-
-//——————————————————————————————————————————————————————————————————————————————
-//   SETUP
-//——————————————————————————————————————————————————————————————————————————————
-
-void can__set(){
-  //--- Switch on builtin led
-  pinMode (LED_BUILTIN, OUTPUT);
-  digitalWrite (LED_BUILTIN, HIGH);
-  
-  //--- Start serial
-  Serial.begin (115200);
-  
-  //--- Wait for serial (blink led at 10 Hz during waiting)
-  while (!Serial) {
-    delay (50);
-    digitalWrite (LED_BUILTIN, !digitalRead (LED_BUILTIN)) ;
-  }
-  
-  //--- There are no default SPI pins so they must be explicitly assigned
-  SPI.setSCK(PICO_SCK_MCP2515);
-  SPI.setTX(PICO_MOSI_MCP2515);
-  SPI.setRX(PICO_MISO_MCP2515);
-  SPI.setCS(PICO_CS_MCP2515);
-  
-  //--- Begin SPI
-  SPI.begin ();
-  
-  //--- Configure ACAN2515
-  Serial.println ("Configure ACAN2515") ;
-  ACAN2515Settings settings (QUARTZ_FREQUENCY, 500UL * 1000UL) ; // CAN bit rate 500s kb/s
-
-  //deleted - no longer needing to be in practice loopback mode
-  //settings.mRequestedMode = ACAN2515Settings::LoopBackMode ; // Select loopback mode
-  
-  const uint16_t errorCode = can.begin (settings, [] { can.isr () ; }) ;
-  if (errorCode == 0) {
-    Serial.print ("Bit Rate prescaler: ") ;
-    Serial.println (settings.mBitRatePrescaler) ;
-    Serial.print ("Propagation Segment: ") ;
-    Serial.println (settings.mPropagationSegment) ;
-    Serial.print ("Phase segment 1: ") ;
-    Serial.println (settings.mPhaseSegment1) ;
-    Serial.print ("Phase segment 2: ") ;
-    Serial.println (settings.mPhaseSegment2) ;
-    Serial.print ("SJW: ") ;
-    Serial.println (settings.mSJW) ;
-    Serial.print ("Triple Sampling: ") ;
-    Serial.println (settings.mTripleSampling ? "yes" : "no") ;
-    Serial.print ("Actual bit rate: ") ;
-    Serial.print (settings.actualBitRate ()) ;
-    Serial.println (" bit/s") ;
-    Serial.print ("Exact bit rate ? ") ;
-    Serial.println (settings.exactBitRate () ? "yes" : "no") ;
-    Serial.print ("Sample point: ") ;
-    Serial.print (settings.samplePointFromBitStart ()) ;
-    Serial.println ("%") ;
-  } else {
-    Serial.print ("Configuration error 0x") ;
-    Serial.println (errorCode, HEX) ;
-  }
- 
-}
-
-static uint32_t gBlinkLedDate = 0 ;
-static uint32_t gReceivedFrameCount = 0 ;
-static uint32_t gSentFrameCount = 0 ;
-
-
-void can__update(){
-   //can.dispatchReceivedMessage();
-  
-  CANMessage frame ;
-
-  /*
-  if (can.available()) {
-    can.receive(frame);
-    Serial.print("ID: ");
-    Serial.println(frame.id);
-    Serial.print("Data: ");
-    for (int i = 0; i < 8; i++) {
-      Serial.println(frame.data[i]);
-    }
-  }
-  */
-  
-  
-  if (gBlinkLedDate < millis ()) {
-    gBlinkLedDate += 2000 ;
-    digitalWrite (LED_BUILTIN, !digitalRead (LED_BUILTIN)) ;
-    const bool ok = can.tryToSend (frame) ;
-    if (ok) {
-      gSentFrameCount += 1 ;
-      Serial.print ("Sent: ") ;
-      Serial.println (gSentFrameCount) ;
-    } else {
-      Serial.println ("Send failure") ;
-    }
-  }
-  if (can.available ()) {
-    can.receive (frame) ;
-    gReceivedFrameCount ++ ;
-    Serial.print ("Received: ") ;
-    Serial.println (gReceivedFrameCount) ;
-  }
-  
-}
