--- can.h
@@ -1,12 +1,0 @@
-#include <ACAN2515.h>
-#include <Arduino.h>
-
-#ifndef CAN_H_
-#define CAN_H_
-
-
-void can__set();
-void can__update();
-
-
-#endif CAN_H_
