+++ icu/leds.h
@@ -1,0 +1,14 @@
+#include <Arduino.h>
+#include <LedControl.h>
+
+#ifndef LEDS_H_
+#define LEDS_H_
+
+
+void leds__wake();
+void leds__start();
+
+//create LEDs Update function that takes CAN data and changes LEDs
+
+
+#endif LEDS_H_
